refactor(http): move src/api/ to src/plugins/http/ (#568)
Pure rename + import-path updates ahead of unified runtime work.
HTTP URLs unchanged. Header comments updated.

Part of issue #568 (unified AgentRuntime).
