+++ src/plugins/http/routes.ts
@@ -1,13 +1,13 @@
-// src/api/routes.ts — Route registry (addRoute / matchRoute)
+// src/plugins/http/routes.ts — Route registry (addRoute / matchRoute)
 
 import type { ServerResponse } from "node:http";
 
-import type { CronScheduler } from "../automation/cron-job.js";
-import type { ConfigReloader } from "../workspace/config-reloader.js";
-import type { DefaultAgentManager } from "../core/agent-manager.js";
-import type { UsageTracker } from "../core/usage-tracker.js";
-import type { SessionStoreManager } from "../core/session-store-manager.js";
-import type { HookRegistry } from "../plugins/hooks.js";
+import type { CronScheduler } from "../../automation/cron-job.js";
+import type { ConfigReloader } from "../../workspace/config-reloader.js";
+import type { DefaultAgentManager } from "../../core/agent-manager.js";
+import type { UsageTracker } from "../../core/usage-tracker.js";
+import type { SessionStoreManager } from "../../core/session-store-manager.js";
+import type { HookRegistry } from "../../plugins/hooks.js";
 import type { ApiRequest } from "./middleware.js";
 
 // ---------------------------------------------------------------------------
